+++ Libraries/XulbuX/CHANGELOG.md
@@ -18,6 +18,7 @@
 ## 28.10.2024 `v1.5.1`
 * Renamed all library files for a better naming convention
 * Now all functions in `xx_color` support both HEX prefixes (`#` and `0x`)
+* Added the default HEX prefix to `_consts_.py`
 * Fixed bug when initializing a `hexa()` object:<br>
   Would throw an error, even if the color was valid